+++ internal/server/transform/transform_clean_header.go
@@ -1,4 +1,4 @@
-package server
+package transform
 
 import (
 	"strings"
