+++ internal/server/transform/transform_max_tokens.go
@@ -1,4 +1,4 @@
-package server
+package transform
 
 import (
 	"github.com/anthropics/anthropic-sdk-go"
